[~] changed came value type int -> float

+++ project/includes/core/reader/IReader.hpp
@@ -20,8 +20,8 @@ namespace raytracer {
         virtual void SetUp() = 0;
 
         virtual std::pair<std::size_t, std::size_t> GetCameraResolution() = 0;
-        virtual std::array<int, 3> GetCameraPosition() = 0;
-        virtual std::array<int, 3> GetCameraRotation() = 0;
+        virtual std::array<float, 3> GetCameraPosition() = 0;
+        virtual std::array<float, 3> GetCameraRotation() = 0;
         virtual double GetCameraFieldOfView() = 0;
 
         virtual std::vector<std::unique_ptr<IObject>> GetObjects() = 0;
