+++ project/includes/core/reader/cfgReader/CfgReader.hpp
@@ -116,8 +116,8 @@ public:
 
     void SetUp();
     std::pair<std::size_t, std::size_t> GetCameraResolution();
-    std::array<int, 3> GetCameraPosition();
-    std::array<int, 3> GetCameraRotation();
+    std::array<float, 3> GetCameraPosition();
+    std::array<float, 3> GetCameraRotation();
     double GetCameraFieldOfView();
 
     std::vector<std::unique_ptr<IObject>> GetObjects() { return loadLists<IObject>("primitives"); };
